MemoristAPI/requirements.txt
boto3==1.9.28
botocore==1.12.28
cfe==0.0.15
Click==7.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-location-field==2.0.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
Pillow==5.3.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
